Refactor progress meter, spinner & log into own package
