+++ commands/command_checkout.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/github/git-lfs/git"
 	"github.com/github/git-lfs/lfs"
+	"github.com/github/git-lfs/progress"
 	"github.com/github/git-lfs/vendor/_nuts/github.com/rubyist/tracerx"
 	"github.com/github/git-lfs/vendor/_nuts/github.com/spf13/cobra"
 )
@@ -116,7 +117,7 @@ func checkoutWithIncludeExclude(include []string, exclude []string) {
 	for _, pointer := range pointers {
 		totalBytes += pointer.Size
 	}
-	progress := lfs.NewProgressMeter(len(pointers), totalBytes, false)
+	progress := progress.NewProgressMeter(len(pointers), totalBytes, false, lfs.Config.Getenv("GIT_LFS_PROGRESS"))
 	progress.Start()
 	totalBytes = 0
 	for _, pointer := range pointers {
