+++ progress/meter.go
@@ -1,11 +1,8 @@
-package lfs
+package progress
 
 import (
 	"fmt"
-	"io"
 	"os"
-	"path/filepath"
-	"runtime"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -37,8 +34,8 @@ type ProgressMeter struct {
 
 // NewProgressMeter creates a new ProgressMeter for the number and size of
 // files given.
-func NewProgressMeter(estFiles int, estBytes int64, dryRun bool) *ProgressMeter {
-	logger, err := newProgressLogger()
+func NewProgressMeter(estFiles int, estBytes int64, dryRun bool, logPath string) *ProgressMeter {
+	logger, err := newProgressLogger(logPath)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "Error creating progress logger: %s\n", err)
 	}
@@ -154,65 +151,6 @@ func (p *ProgressMeter) update() {
 	fmt.Fprintf(os.Stdout, out)
 }
 
-// progressLogger provides a wrapper around an os.File that can either
-// write to the file or ignore all writes completely.
-type progressLogger struct {
-	writeData bool
-	log       *os.File
-}
-
-// Write will write to the file and perform a Sync() if writing succeeds.
-func (l *progressLogger) Write(b []byte) error {
-	if l.writeData {
-		if _, err := l.log.Write(b); err != nil {
-			return err
-		}
-		return l.log.Sync()
-	}
-	return nil
-}
-
-// Close will call Close() on the underlying file
-func (l *progressLogger) Close() error {
-	if l.log != nil {
-		return l.log.Close()
-	}
-	return nil
-}
-
-// Shutdown will cause the logger to ignore any further writes. It should
-// be used when writing causes an error.
-func (l *progressLogger) Shutdown() {
-	l.writeData = false
-}
-
-// newProgressLogger creates a progressLogger based on the presence of
-// the GIT_LFS_PROGRESS environment variable. If it is present and a log file
-// is able to be created, the logger will write to the file. If it is absent,
-// or there is an err creating the file, the logger will ignore all writes.
-func newProgressLogger() (*progressLogger, error) {
-	logPath := Config.Getenv("GIT_LFS_PROGRESS")
-
-	if len(logPath) == 0 {
-		return &progressLogger{}, nil
-	}
-	if !filepath.IsAbs(logPath) {
-		return &progressLogger{}, fmt.Errorf("GIT_LFS_PROGRESS must be an absolute path")
-	}
-
-	cbDir := filepath.Dir(logPath)
-	if err := os.MkdirAll(cbDir, 0755); err != nil {
-		return &progressLogger{}, err
-	}
-
-	file, err := os.OpenFile(logPath, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0666)
-	if err != nil {
-		return &progressLogger{}, err
-	}
-
-	return &progressLogger{true, file}, nil
-}
-
 func formatBytes(i int64) string {
 	switch {
 	case i > 1099511627776:
@@ -226,59 +164,4 @@ func formatBytes(i int64) string {
 	}
 
 	return fmt.Sprintf("%d B", i)
-}
-
-// Indeterminate progress indicator 'spinner'
-type Spinner struct {
-	stage int
-	msg   string
-}
-
-var spinnerChars = []byte{'|', '/', '-', '\\'}
-
-// Print a spinner (stage) to out followed by msg (no linefeed)
-func (s *Spinner) Print(out io.Writer, msg string) {
-	s.msg = msg
-	s.Spin(out)
-}
-
-// Just spin the spinner one more notch & use the last message
-func (s *Spinner) Spin(out io.Writer) {
-	s.stage = (s.stage + 1) % len(spinnerChars)
-	s.update(out, string(spinnerChars[s.stage]), s.msg)
-}
-
-// Finish the spinner with a completion message & newline
-func (s *Spinner) Finish(out io.Writer, finishMsg string) {
-	s.msg = finishMsg
-	s.stage = 0
-	var sym string
-	if runtime.GOOS == "windows" {
-		// Windows console sucks, can't do nice check mark except in ConEmu (not cmd or git bash)
-		// So play it safe & boring
-		sym = "*"
-	} else {
-		sym = fmt.Sprintf("%c", '\u2714')
-	}
-	s.update(out, sym, finishMsg)
-	out.Write([]byte{'\n'})
-}
-
-func (s *Spinner) update(out io.Writer, prefix, msg string) {
-
-	str := fmt.Sprintf("%v %v", prefix, msg)
-
-	width := 80 // default to 80 chars wide if ts.GetSize() fails
-	size, err := ts.GetSize()
-	if err == nil {
-		width = size.Col()
-	}
-	padding := strings.Repeat(" ", width-len(str))
-
-	fmt.Fprintf(out, "\r%v%v", str, padding)
-
-}
-
-func NewSpinner() *Spinner {
-	return &Spinner{}
 }
